requirements.txt
Django==2.2.3
django-filter==2.2.0
djangorestframework==3.10.1
pytz==2019.1
sqlparse==0.3.0
